move bulk of notification command logic to abstract class
